Adjust buffer sizes on ia64.
From the debian change entry:

> Finish fixing the ia64 buffer issue. Previous commits increased
YY_READ_BUF_SIZE (where __ia64__ is defined) but left YY_BUF_SIZE
unchanged, so that didn't fix the problem in the end. In the general
case, the latter is twice the former. Therefore set it to the same
ratio in the ia64 case.

In general, this sort of architecture specific fix is not the path we
want to take, but the cleanup should be done in a more organized way
in the future and getting it working would be preferrable now.