Replaced CHAR macro with unsigned char type.
Thanks to Michael McConville for pointing out that the old Char macro causes problems with static analysis.  The macro has been removed and replaced with 'unsigned char' throughout the flex sources. The macro is not needed at best and was confusing at worst. It was not used in any of the example files nor was it mentioned in the manual at all.