Remove flex_free(), a free() alias
Same deal as flex_alloc() and flex_realloc(): this is an alias for a
stdlib function and no longer needs to exist.

Also, buf_destroy()'s logic doesn't make sense. It checks whether buf is
NULL in one block and unconditionally dereferences it later. I think the
right thing to do here is make buf_destroy() NULL-safe, which,
incidentally, is what I did.