Ditch libcompat.la. Link $(LIBOBJS) from src/ dir.
This helps bootstrapping flex as we no longer need AR or RANLIB for
such a little library.

All of its source files (malloc.c realloc.c reallocarray.c) are now
linked individually from the src/ directory.

By the way, the use of libtool on libcompat was not even recommended in
Automake manual [1], before this ditching.

[1] https://www.gnu.org/software/automake/manual/html_node/LIBOBJS.html
   "Because '$(LIBOBJS)' and '$(ALLOCA)' contain object file names that
    end with '.$(OBJEXT)', they are not suitable for Libtool libraries
    (where the expected object extension is .lo): LTLIBOBJS and LTALLOCA
    should be used instead."