test: Never unlist pthread test; add "skip" status for it.
"Unlisting" the pthread test when libpthread is not available is never
a good idea. Should let it compile into a stub program that returns the
"skip" status.

configure.ac and tests/Makefile.am are edited so that pthread test can
be built anyway, but only works (not as the stub) when both pthread.h
and libpthread are detected on the system.

Also fix a small typo in tests/pthread.l comments.