scanner: free strings if allocated on heap
If the '%option yyclass' is specified more than once, a memory leak
occurs. Additionally, --yyclass=STRING will set yyclass to the argument,
which should not be freed.

The same occurs with outfile, header-file and prefix. Also, prefix is
initially assigned the const string "yy".

The memory leak is fixed by freeing the pointer (e.g. yyclass) if
previously allocated. The issue with freeing an argument or const string
if fixed by always assigning these variables with a strdup().