No need to generate lib/Makefile.in
Files in lib/ are picked up and built using using makefile directives
in src/Makefile.am. Remove the need to generate lib/Makefile.in and the
dummy lib/Makefile.am.

(I discoved this tweak on my attempt to change flex to use "non-
recursive makefile". I'm not sure if flex should be migrated to use
non-recursive makefile, but it's good to get rid of lib/Makefile.in
first.)