Change rule_has_nl[] and ccl_has_nl[] allocation.
Also remove reallocate_bool_array() and allocate_bool_array() as these
are no longer needed.

(I intend that all arrays of boolean values be stored as either
bitfields or char arrays. _Bool may be defined with size of int for
some compilers - not a good idea for an array type.)