scan.l: fixing bug that flex ignores start condition  name directly following '%s' or '%x'.
flex 2.6.4 ignores the start conditions A and B in the
following sample:

~~~c
  /* sample reproducing the bug */

%sA A1
%xB B1

%%
.

  /* EOF */
~~~

Backward compatibility is honored, albeit code in the past
should not have defined any unused start condition which
immediately follows %s or %x.

Sample test added to the test suite.

Clarification added to the manual that, after the declarator
'%s' or '%x' at BOL, flex expects a non-void list of start
condition names until EOL. As usual, multiple lines with
declarations are possible.