build: Generate deterministic scan.c and stage1scan.c.
Change the generating rules of scan.c (in dist-hook) and stage1scan.c
so that the source and output file names in #line directives are fixed.
This would allow the scanner results be comparable to each other.

Also in --disable-bootstrap build, the flex binary will now be built
with scan.c directly - no more sedding.

[v4: Don't leave an incomplete stage1scan.c file if stage1flex fails
for any reason. This ensures the file is regenerated when `make` is
invoked again (2020-03-09)]

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>