Remove guards on num_backing_up in the code generation,
These tun out to be redundant because num_backing_up is
incremented at the begining of gentabs.

The verification that this change is correct is not just that the test
suite passes, but that generation of code across the test suite is
not perturbed in any way when these guards are removed.

The variable num_backing_up still exists but is now only used to count
DFA backups for reporting purposes.

There was actually a code defect here.  The increment in gentabs
was bogus and probably a thoughtless hack applied when table
compression was incremented.