Begin replacing method table entries with hook macros
Do this for table opener/closer/continuation syntax, the trace-format
string, the state entry string, constant definitions, the state-dyad
format, and the three pieces of EOF state syntax.  The documentation
appendix on how to write a back end is also updated.

There are comment diffs because I decided generating an
explicit fallthrough marker and some other new explanatory comments
was a good idea.

#58 in the retargeting patch series