Macroize all the remaining code-generation methods.
Produces a tediously large diff in generated test code that
is all table moving around.  This is due to them being shipped
as macros and being substitured in a fixed order determined
by the calls in the skell file, rather than veing generated as
the functrions originally emitting the tables are called.

#68 in the retargeting patch series