Finish up macro abstraction.
Everything Flex ships to the skeleton-file expansion phase is
now either a macro expansion or a macro call.  This almost finishes
the retargeting patch series; the wrapup will be documentation.

Sadky, this does not get us *all* the way to target-syntax independence.
The probem is the inclusion of tables_shared.c when table serialization is
enabled. Which ,eans table serialization is not practical to support
outside the C/C++ back end.

No diffs in generated test code from this commit.

#69 in the retargeting patch series