First step towards an interface independent of C macros.
This patch implements and documents a yyreject() macro to replace
argumentless REJECT.  It does not remove REJECT, but warns that this
macro will not be supported in non-C languages and deprecates it.

This commit begins a new appendix in the Flex manual, to list
deprecated interfaces and explain why they have been superseded.