Rename unput() to yyunput(); follow through in option switches.
This change was not strictly necessary to sever a preprocessor
dependency nor make the API uniform aross both C and C++.  But it
cried out to be made, because now *all* the rule hooks are in the yy*
namespace.  This makes the API easier to document and remember.

unput() is left in place as a compatibility macro for existing
users, but only documented as a legacy interface.  The "unput"
variants of switches and options have also been retained.