Begin work on a C99 front end.
The idea behind this is to exploit the existing test suite - write a
stripped-down, cleaned up C skeleton that only does reentrant scanners
and tosses out features like loadable tables and header generation
that can't be ported to other languages.

The C99 skeleton can then serve both as a launch point for the Go back
end and a model for back ends in procedural languages without OO.  The
initial version is identical to the cpp skeleton except for different
name and suffix properties abd having M4_MODE_C_OLY forcrd on.  The
latter means it blows up on the C++ tests.