Another big sinmplification of the test machinery.
If you drop a file with the extension ".rules" in tests/,
"make check" will now automatically make and run tests
for all back ends - presently nr, r, and c99.

It would be trivial to change ruleset.sh so all table-compression
options are checked for all rulesets, and probably a good idea.