In C99, remove most code dependencies on yyin/yyout/yyleng/yytext macros.
These served a purpose when the code had to handle both reentrant and
non-reentrant cases.  Now, while they should still be available in user
actions, the generated code is much less confusing if the yyg pointer
in the references is explicit.

Also, drop VMS support and consolidate platform-specific references
at one spot. This is in the same commit with the macro cleanup because
it included getting rid of the last stray yyin/yyout references.