YY_AT_BOL() -> yyatbol(), yy_set_bol() -> yysetbol(), and make them magic.
The tradeoff is a bit delicate here; yy_set_bol() did not strictly
speaking need to change. On the other hand, squeezing the underscores
out of YY_AT_BOL() but not yy_set_bol() would have been exactly the
kind of irregularity that people trip over and curse.

The underlying issue is of course languages like Go and Python with
validators that get upset by all caps and embedded underscores. On the
one hand, we want to not be churning the Flex API gratuitously.  On
the other hand it is desirable if as few entry point names as possible
need to change in different target languages.

The approach I'm taking to deciding this question in individual cases
is this: if an entry point name is likely to appear in actions, and
must be recognized as magic by Flex itself in order to have a final
context argument added or an object reference prepended, then it must
be multilanguage-safe - not all-caps and no underscores.

Any name that does *not* need to be wired into Flex can be left
alone. Users may have to change it when they port.