Fix typo revealed by Go porting work, and improve test coverage to match.
We a ruleset test for BOL matching and a test for the array case of yymore.
Improve yymore testing by adding tests that require a BOL match in
both array and pointer cases.

Also, convert lineno sporadic test to be ruleset-based so it's
applied on all back ends.

No cide changes.