build: Generate deterministic scan.c and stage1scan.c.
Change the generating rules of scan.c (in dist-hook) and stage1scan.c
so that the source and output file names in #line directives are fixed.
This would allow the scanner results be comparable to each other.

Also in --disable-bootstrap build, the flex binary will now be built
with scan.c directly - no more sedding.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>